Python script required by YouCompleteMe - template
Added the Python script required by YouCompleteMe to point to the
relevant virtual environment. Or rather, a template for that file. The
most obvious way to use it is to create a symlink to it at the root of
the project. But actually using the file in this commit as a template by
copying it under the project root and adapting it works too.  A file
named .ycm_extra_conf.py at the root of the project will not be under
revision control.